+++ minijinja/src/lib.rs
@@ -137,6 +137,7 @@ mod vm;
 
 pub mod filters;
 pub mod functions;
+pub mod meta;
 pub mod syntax;
 pub mod tests;
 pub mod value;